Only write chart.l10n.in if it doesn't exist already
So that 'helm l10n init' can be safely used on existing l10n projects to
update templates

+++ helm-l10n.sh
@@ -33,7 +33,8 @@ function init()
     sed "s/CHARTNAME/$chart_name/g" $HELM_PLUGIN_DIR/l10n.yaml > $chart/templates/l10n.yaml
     sed "s/CHARTNAME/$chart_name/g" $HELM_PLUGIN_DIR/l10n-subcharts.yaml > $chart/templates/l10n-subcharts.yaml
     mkdir $chart/po 2> /dev/null || :
-    cat > $chart/chart.l10n.in <<-EOF
+    if [ ! -f "$chart/chart.l10n.in" ]; then
+        cat > $chart/chart.l10n.in <<-EOF
 	Name=$chart_name
 	Instance={{ .Release.Name }}
 
@@ -43,6 +44,7 @@ function init()
 	#[bahman:postgres]
 	#Instance=Bahman subchart of postgres
 	EOF
+    fi
     echo "Basic l10n structure created. You can now add source strings to $chart/chart.l10n.in"
 }
